Move service classes to a dedicated folder
