+++ Elecritic/Pages/Categories_cellphones.razor.cs
@@ -2,6 +2,7 @@
 using System.Threading.Tasks;
 
 using Elecritic.Models;
+using Elecritic.Services;
 
 using Microsoft.AspNetCore.Components;
