+++ Elecritic/Pages/ProductPage.razor.cs
@@ -3,6 +3,7 @@ using System.ComponentModel.DataAnnotations;
 using System.Threading.Tasks;
 
 using Elecritic.Models;
+using Elecritic.Services;
 
 using Microsoft.AspNetCore.Components;
